+++ src/CustomType/Data/Field/BooleanField.php
@@ -2,6 +2,8 @@
 
 namespace Torr\PrismicApi\CustomType\Data\Field;
 
+use Torr\PrismicApi\CustomType\Helper\FilterFieldsHelper;
+
 /**
  * @see https://prismic.io/docs/core-concepts/boolean
  */
@@ -20,7 +22,7 @@ final class BooleanField extends InputField
 		?string $placeholderTrue = null,
 	)
 	{
-		parent::__construct(self::TYPE_KEY, $this->filterOptionalFields([
+		parent::__construct(self::TYPE_KEY, FilterFieldsHelper::filterOptionalFields([
 			"label" => $label,
 			"placeholder_false" => $placeholderFalse,
 			"placeholder_true" => $placeholderTrue,