+++ src/CustomType/Data/Field/ImageField.php
@@ -4,6 +4,7 @@ namespace Torr\PrismicApi\CustomType\Data\Field;
 
 use Torr\PrismicApi\CustomType\Data\Part\ImageConstraint;
 use Torr\PrismicApi\CustomType\Data\Part\Thumbnail;
+use Torr\PrismicApi\CustomType\Helper\FilterFieldsHelper;
 
 /**
  * @see https://prismic.io/docs/core-concepts/image
@@ -25,7 +26,7 @@ final class ImageField extends InputField
 		array $thumbnails = [],
 	)
 	{
-		parent::__construct(self::TYPE_KEY, $this->filterOptionalFields([
+		parent::__construct(self::TYPE_KEY, FilterFieldsHelper::filterOptionalFields([
 			"label" => $label,
 			"placeholder" => $placeholder,
 			"constraint" => $imageConstraint?->toArray(),